requirements.txt
aiogram~=3.17.0
moviepy~=1.0.3
python-dotenv~=1.0.0